remove unnecessary boxing + slight rename
