+++ kube/src/lib.rs
@@ -181,7 +181,7 @@ pub use kube_core as core;
 // Mock tests for the runtime
 #[cfg(test)]
 #[cfg(all(feature = "derive", feature = "runtime"))]
-mod mocks;
+mod mock_tests;
 
 // Tests that require a cluster and the complete feature set
 // Can be run with `cargo test -p kube --lib --features=runtime,derive -- --ignored`
